create formatAttachmentUrl file
