+++ app/lib/utils.ts
@@ -1,15 +1,5 @@
 import { coerce, gt, gte, lt, lte, SemVer } from 'semver';
 
-export const formatAttachmentUrl = (attachmentUrl: string | undefined, userId: string, token: string, server: string): string => {
-	if (attachmentUrl && attachmentUrl.startsWith('http')) {
-		if (attachmentUrl.includes('rc_token')) {
-			return encodeURI(attachmentUrl);
-		}
-		return encodeURI(`${attachmentUrl}?rc_uid=${userId}&rc_token=${token}`);
-	}
-	return encodeURI(`${server}${attachmentUrl}?rc_uid=${userId}&rc_token=${token}`);
-};
-
 const methods = {
 	lowerThan: lt,
 	lowerThanOrEqualTo: lte,
